requirements-pip2.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.10.0
blinker==1.4
Brotli==1.0.9
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
colorful==0.5.4
cryptography==36.0.1
cycler==0.11.0
dash==2.0.0
dataclasses==0.8
debugpy==1.5.1
decorator==5.1.0
executing==0.8.2
Flask==2.0.2
Flask-Compress==0.0.0
fonttools==4.28.5
frozenlist==1.2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==2.10.0
hdbscan==0.8.27
hiplot==0.1.32
icecream==2.1.1
idna==3.1
imageio==2.13.5
importlib-metadata==4.10.0
importlib-resources==5.4.0
ipykernel==6.0.3
ipython==7.30.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.3
jupyter-client==6.1.12
jupyter-core==4.9.1
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
llvmlite==0.36.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mpld3==0.5.7
multidict==5.2.0
munkres==1.1.4
nbformat==5.1.3
networkx==2.6.3
numba==0.53.0
numpy==1.22.0
oauthlib==3.1.1
olefile==0.46
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
patsy==0.5.2
pickleshare==0.7.5
Pillow==8.4.0
pip==21.3.1
plotly==5.5.0
prettyprinter==0.18.0
prompt-toolkit==3.0.24
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
Pygments==2.11.1
PyJWT==2.3.0
pynndescent==0.5.5
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyreadline==2.1
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pyu2f==0.1.5
pywin32==303
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
setuptools==60.2.0
sip==4.19.13
six==1.16.0
soupsieve==2.3.1
statsmodels==0.13.1
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==3.0.0
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
umap-learn==0.5.2
unicodedata2==14.0.0
urllib3==1.26.7
watermark==2.3.0
wcwidth==0.2.5
Werkzeug==2.0.1
wheel==0.37.1
win-inet-pton==1.1.0
wrapt==1.13.3
yarl==1.7.2
zipp==3.6.0
